+++ libsolidity/formal/SymbolicTypes.cpp
@@ -57,8 +57,13 @@ smt::SortPointer dev::solidity::smtSort(Type const& _type)
 			solAssert(mapType, "");
 			return make_shared<smt::ArraySort>(smtSort(*mapType->keyType()), smtSort(*mapType->valueType()));
 		}
-		// TODO Solidity array
-		return make_shared<smt::Sort>(smt::Kind::Int);
+		else
+		{
+			solAssert(isArray(_type.category()), "");
+			auto arrayType = dynamic_cast<ArrayType const*>(&_type);
+			solAssert(arrayType, "");
+			return make_shared<smt::ArraySort>(make_shared<smt::Sort>(smt::Kind::Int), smtSort(*arrayType->baseType()));
+		}
 	}
 	default:
 		// Abstract case.
@@ -82,7 +87,7 @@ smt::Kind dev::solidity::smtKind(Type::Category _category)
 		return smt::Kind::Bool;
 	else if (isFunction(_category))
 		return smt::Kind::Function;
-	else if (isMapping(_category))
+	else if (isMapping(_category) || isArray(_category))
 		return smt::Kind::Array;
 	// Abstract case.
 	return smt::Kind::Int;
@@ -92,7 +97,8 @@ bool dev::solidity::isSupportedType(Type::Category _category)
 {
 	return isNumber(_category) ||
 		isBool(_category) ||
-		isMapping(_category);
+		isMapping(_category) ||
+		isArray(_category);
 }
 
 bool dev::solidity::isSupportedTypeDeclaration(Type::Category _category)
@@ -140,6 +146,8 @@ pair<bool, shared_ptr<SymbolicVariable>> dev::solidity::newSymbolicVariable(
 	}
 	else if (isMapping(_type.category()))
 		var = make_shared<SymbolicMappingVariable>(type, _uniqueName, _solver);
+	else if (isArray(_type.category()))
+		var = make_shared<SymbolicArrayVariable>(type, _uniqueName, _solver);
 	else
 		solAssert(false, "");
 	return make_pair(abstract, var);
@@ -196,6 +204,11 @@ bool dev::solidity::isFunction(Type::Category _category)
 bool dev::solidity::isMapping(Type::Category _category)
 {
 	return _category == Type::Category::Mapping;
+}
+
+bool dev::solidity::isArray(Type::Category _category)
+{
+	return _category == Type::Category::Array;
 }
 
 smt::Expression dev::solidity::minValue(IntegerType const& _type)